+++ mkspec/gxx_mkspecs.py
@@ -197,6 +197,24 @@ def cxx_gxx49_x86(conf):
 
 
 @conf
+def cxx_gxx52_x64(conf):
+    """
+    Detect and setup the g++ 4.9 compiler for 64 bit
+    """
+    conf.mkspec_gxx_configure(5, 2)
+    conf.mkspec_add_common_flag('-m64')
+
+
+@conf
+def cxx_gxx52_x86(conf):
+    """
+    Detect and setup the g++ 5.2 compiler for 32 bit
+    """
+    conf.mkspec_gxx_configure(5, 2)
+    conf.mkspec_add_common_flag('-m32')
+
+
+@conf
 def cxx_raspberry_gxx47_arm(conf):
     """
     Detect and setup the g++ 4.7 cross-compiler for Raspberry Pi (Linux)